Divided form into several components
